fix: only put resolved authors in es  * ref https://github.com/cern-sis/issues-inspire/issues/970
